put back in different files
https://coveord.atlassian.net/browse/KIT-1725
